feat(frontend): more cleanups
