+++ packages/frontend/src/components/UserApps.vue
@@ -1,6 +1,9 @@
 <template>
-  <v-card class="elevation-0 mt-3 mb-5">
-    <v-card-title>Applications</v-card-title>
+  <v-card rounded="lg" :class="`${!$vuetify.theme.dark ? 'grey lighten-5' : ''}`">
+    <v-toolbar flat>
+      <v-toolbar-title>Applications</v-toolbar-title>
+    </v-toolbar>
+
     <v-card-text>
       Register and manage third-party Speckle Apps that, once authorised by a user on this server,
       can act on their behalf.
