+++ packages/frontend/src/components/admin/VersionInfoCard.vue
@@ -52,7 +52,6 @@ export default {
         query {
           serverInfo {
             version
-            isDevServer
           }
         }
       `,
